test_completion_cost_deepseek
